Reworked all measure types to use LayoutLength
Using Child Properties for layout parameters means that all values have to be
Property::INTEGER type; i.e. int. Have removed all uint16_t types, all uint32_t
types, all Uint16Pair types, and replaced with LayoutLength or LayoutLength::IntType.

Desired size constants of MATCH_PARENT or WRAP_CONTENT are now negative values in
the same integer type.

Updated tree dumper (partially in tests, partially in control-data-impl-debug) to
output valid JSON.

+++ dali-toolkit/devel-api/layouting/child-layout-data.h
@@ -23,11 +23,11 @@ namespace Dali
 namespace Toolkit
 {
 
+// used in property system - must be integer values.
 namespace ChildLayoutData
 {
-
-static const uint16_t MATCH_PARENT = uint16_t(-1); ///< Constant which indicates child size should match parent size
-static const uint16_t WRAP_CONTENT = uint16_t(-2); ///< Constant which indicates parent should take the smallest size possible to wrap it's children with their desired size
+static const int MATCH_PARENT = -1; ///< Constant which indicates child size should match parent size
+static const int WRAP_CONTENT = -2; ///< Constant which indicates parent should take the smallest size possible to wrap it's children with their desired size
 
 } // namespace ChildLayoutData
